Little bug fix for implicit cast that caused a compilation error with Ubuntu.
svn path=/trunk/iCub/; revision=21341

+++ src/modules/skinManager/src/compensator.cpp
@@ -325,7 +325,7 @@ bool Compensator::readRawAndWriteCompensatedData(){
         // we want to use the filtered values
         touchDetectedFilt[i] = (d > touchThresholds[i] + addThreshold);
         if(binarization)
-		    d = touchDetectedFilt[i] ? BIN_TOUCH : BIN_NO_TOUCH;
+		    d = (double)(touchDetectedFilt[i] ? BIN_TOUCH : BIN_NO_TOUCH);
         
 	    compensatedData2Send[i] = d;
 	}
